feat: create the fully runnable vision model
- fix model value check for complex param (replace complex with normal in future)
- create ByteLatentTransformer vision version with debug logs
- adapt training script to the change
- refactor local models new modules (ImageEncoder, ImageDecoder and PredictionHead) and add init_weights for them

+++ bytelatent/data/shape_data_stream.py
@@ -176,7 +176,7 @@ class ShapeDataset:
             self.current_steps += 1
 
             yield VisionBatch(
-                x=torch.stack(batch_images_x),
+                x=torch.stack(batch_images_x).to(torch.float32),
                 y=torch.stack(batch_images_y),
                 batch_t=torch.arange(start=1, end=current_batch_time + 1).to(device=self.device),
                 angles=torch.stack(batch_angles).to(device=self.device),
